rename errors module to error
