+++ zebra-consensus/src/error.rs
